apps/account/views/viewset.py
from rest_framework import viewsets, permissions

from apps.account.models import Student, Teacher, PrivateMessage
from apps.account.serializers import StudentSerializer, TeacherSerializer, PrivateMessageSerializer

class StudentViewSet(viewsets.ModelViewSet):
    queryset = Student.objects.all()
    serializer_class = StudentSerializer


class TeacherViewSet(viewsets.ModelViewSet):
    queryset = Teacher.objects.all()
    serializer_class = TeacherSerializer


class PrivateMessageViewSet(viewsets.ModelViewSet):
    queryset = PrivateMessage.objects.all()
    serializer_class = PrivateMessageSerializer