requirements.txt
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
Django==2.2
mysqlclient==2.1.0
Pillow==9.1.0
pip==20.1.1
pytz==2022.1
setuptools==47.1.1
sqlparse==0.4.2
wheel==0.34.2

djangorestframework~=3.13.1